fix an issue

+++ IdentityCore/src/telemetry/request_telemetry/MSIDLastRequestTelemetry.m
@@ -280,7 +280,7 @@ static int maxErrorCountToArchive = 75;
     if (saveLocation)
     {
         // Some testing has determined that 75 errors corresponds to an archive size of about 8kb. 
-        if (_errorsInfo.count > maxErrorCountToArchive)
+        if ((int)_errorsInfo.count > maxErrorCountToArchive)
         {
             MSID_LOG_WITH_CTX_PII(MSIDLogLevelVerbose, nil, @"Telemetry size over limit when saving to disk, cutting down to limit", nil);
